test: refactor token blacklist test for speed
